requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
pytest-mock>=3.14.0
black>=24.0.0
